+++ frontend/src/components/Sidebar/index.tsx
@@ -1,23 +1,8 @@
 import { Link } from 'react-router-dom';
 import OptionItem from './OptionItem';
 import OptionList from './OptionList';
-import { useState } from 'react';
-
-export type ArrowDirection = 'right' | 'down' | 'up';
-
-export type OptionDecoration = {
-  fontSize?: string;
-  fontWeight?: string;
-  underlineHeight?: string;
-  direction?: ArrowDirection | null;
-  disabled?: boolean;
-};
-
-type Option = {
-  name: string;
-  link: string;
-  isGetReady?: boolean;
-} & OptionDecoration;
+import { Option } from '../../types';
+import Accordion from '../Accordion';
 
 export type SidebarProps = {
   onSidebarOpen: (isOpen: boolean) => void;
@@ -109,12 +94,6 @@ const Sidebar = ({ onSidebarOpen, isOpen }: SidebarProps) => {
     },
   ];
 
-  const [helpClicked, setHelpClicked] = useState(false);
-
-  const toggleHelp = () => {
-    return setHelpClicked(!helpClicked);
-  };
-
   return (
     <OptionList onSidebarOpen={onSidebarOpen} isOpen={isOpen}>
       {userSetting.map((option) => (
@@ -132,46 +111,7 @@ const Sidebar = ({ onSidebarOpen, isOpen }: SidebarProps) => {
         </Link>
       ))}
 
-      <div id="accordion">
-        <div id="고객센터" onClick={toggleHelp}>
-          {help.title.map((option) => {
-            const dynamicDirection = helpClicked ? 'up' : 'down';
-
-            return (
-              <OptionItem
-                key={option.name}
-                fontSize={option.fontSize}
-                fontWeight={option.fontWeight}
-                underlineHeight={option.underlineHeight}
-                direction={dynamicDirection}
-                disabled={option.disabled}
-                isGetReady={option.isGetReady}
-              >
-                {option.name}
-              </OptionItem>
-            );
-          })}
-        </div>
-
-        {helpClicked && (
-          <div id="옵션">
-            {help.option.map((option) => (
-              <Link key={`${option.name}-${option.link}`} to={option.link}>
-                <OptionItem
-                  fontSize={option.fontSize}
-                  fontWeight={option.fontWeight}
-                  underlineHeight={option.underlineHeight}
-                  direction={option.direction}
-                  disabled={option.disabled}
-                  isGetReady={option.isGetReady}
-                >
-                  {option.name}
-                </OptionItem>
-              </Link>
-            ))}
-          </div>
-        )}
-      </div>
+      <Accordion options={help} />
 
       {connection.map((option) => (
         <Link key={`${option.name}-${option.link}`} to={option.link}>
