+++ frontend/src/types/index.ts
@@ -23,3 +23,19 @@ export const CARD_WIDTH = 69.06;
 export const CARD_HEIGHT = CARD_WIDTH * 0.7451;
 
 export type slidDirection = 'left' | 'right';
+
+export type ArrowDirection = 'right' | 'down' | 'up';
+
+export type OptionDecoration = {
+  fontSize?: string;
+  fontWeight?: string;
+  underlineHeight?: string;
+  direction?: ArrowDirection | null;
+  disabled?: boolean;
+};
+
+export type Option = {
+  name: string;
+  link: string;
+  isGetReady?: boolean;
+} & OptionDecoration;
